+++ src/bed.h
@@ -8,6 +8,6 @@ int is_bed_file(uint8_t *bed);
 
 int has_valid_dimensions(size_t length, int nrows, int ncols);
 
-int extract_genotype(uint8_t *bed, int n, int k, int na_value);
+int extract_genotype(uint8_t *bed, int n, size_t k, int na_value);
 
 #endif
